Time: 0 ms (100.00%) | Memory: 42.1 MB (68.39%) - LeetSync

+++ 70-climbing-stairs/climbing-stairs.java
@@ -1,10 +1,10 @@
 class Solution {
     public int climbStairs(int n) {
-        int f = (int)(Math.round(
+        long f = Math.round(
     (Math.pow((1 + Math.sqrt(5)) / 2, n + 1)
    - Math.pow((1 - Math.sqrt(5)) / 2, n + 1))
    / Math.sqrt(5)
-));
-     return f;
+);
+     return (int)f;
     }
 }
